JavaEE - Five Mapping Types
