+++ .idea/compiler.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="CompilerConfiguration">
+    <annotationProcessing>
+      <profile name="Maven default annotation processors profile" enabled="true">
+        <sourceOutputDir name="target/generated-sources/annotations" />
+        <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
+        <outputRelativeToContentRoot value="true" />
+        <module name="Empty_String_Mapping" />
+        <module name="Exact_Mapping" />
+        <module name="Default_Mapping" />
+        <module name="Extension_Mapping" />
+        <module name="Wild_Card_Mapping" />
+      </profile>
+    </annotationProcessing>
+  </component>
+</project>
