+++ 4_Wild_Card_Mapping/src/main/java/lk/ijse/aad/gdse68/wild_card_mapping/Demo.java
@@ -1,0 +1,30 @@
+package lk.ijse.aad.gdse68.wild_card_mapping;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
+
+@WebServlet(urlPatterns = "/demo", loadOnStartup = 2)
+public class Demo extends HttpServlet {
+
+    @Override
+    public void init() throws ServletException {
+        System.out.println("Hello Init");
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        resp.getWriter().println("Hello Get Method");
+        System.out.println(Thread.currentThread().getName()); //different threads for different requests
+        System.out.println("Hello Get Method");
+    }
+
+    @Override
+    public void destroy() { //
+        System.out.println("Destroyed..........");
+    }
+}
