--- src/main/webapp/WEB-INF/web.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns="https://jakarta.ee/xml/ns/jakartaee"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee https://jakarta.ee/xml/ns/jakartaee/web-app_6_0.xsd"
-         version="6.0">
-
-<!--    <servlet-mapping>-->
-<!--        <servlet-name>Demo</servlet-name>-->
-<!--        <url-pattern>/demo</url-pattern>-->
-<!--    </servlet-mapping>-->
-
-<!--    <servlet>-->
-<!--        <servlet-name>Demo</servlet-name>-->
-<!--        <servlet-class>lk.ijse.aad.gdse68.jakarthaee_intro.Demo</servlet-class>-->
-<!--        <load-on-startup>5</load-on-startup>-->
-<!--    </servlet>-->
-</web-app>
